lib.rs is now empty
* Collections and PtObjects go to model.rs
* PtObject renamed Model
* ctx_from_path! no more exported
