+++ src/common_format.rs
@@ -1,3 +1,19 @@
+// Copyright 2017-2018 Kisio Digital and/or its affiliates.
+//
+// This program is free software: you can redistribute it and/or
+// modify it under the terms of the GNU General Public License as
+// published by the Free Software Foundation, either version 3 of the
+// License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful, but
+// WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+// General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see
+// <http://www.gnu.org/licenses/>.
+
 use Result;
 use failure::ResultExt;
 use csv;
@@ -5,7 +21,7 @@ use std::path;
 use collection::*;
 use objects::{Calendar, Date, ExceptionType};
 use utils::{de_from_date_string, ser_from_naive_date};
-use Collections;
+use model::Collections;
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct CalendarDate {
