+++ src/utils.rs
@@ -72,7 +72,6 @@ where
     })
 }
 
-#[macro_export]
 macro_rules! ctx_from_path {
     ($path: expr) => {
         |_| format!("Error reading {:?}", $path)
